+++ src/util/current-full-year-factory.ts
@@ -1,3 +1,0 @@
-export function currentFullYearFactory(): string {
-	return new Date().getFullYear().toString();
-}