+++ src/util/now-month-factory.ts
@@ -1,0 +1,3 @@
+export function nowMonthFactory(): string {
+	return (new Date().getMonth() + 1).toString();
+}