+++ src/util/now-year-factory.ts
@@ -1,0 +1,3 @@
+export function nowYearFactory(): string {
+	return new Date().getFullYear().toString();
+}