+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Upcoming
 
+- Feature: Support standard semantic version bumps "preminor", "prepatch", "premajor"
+
+- Feature: Support novel semantic version version bump  "date" and "predate" for date-based versioning e.g. "2022.11.0"
+
 ## carnesen-dev-0.0.0 (2022-02-25)
 
 Initial release